integrated the customer, employee and sales

+++ server/controllers/returnBillControllers.js
@@ -32,10 +32,16 @@ const createReturn=expressAsyncHandler(async(req,res)=>
 });
 
 const deleteReturnBill=expressAsyncHandler(async(req,res)=>{
+    const permission=req.user.isAdmin;
     const _id=req.body;
-
+    
+    if(permission)
+    {
     const bill=await ReturnBill.findByIdAndDelete({_id});
     res.status(200).json(bill);
+    }
+    else
+    res.status(400).json("You are not authorized to delete Bills!");
 })
 
 export {deleteReturnBill, createReturn};
